[rename] #6 - dto 폴더명 변경 (chatbot->chat)
