+++ src/main/java/com/BioTy/Planty/dto/chat/SendMessageRequestDto.java
@@ -1,4 +1,4 @@
-package com.BioTy.Planty.dto.chatbot;
+package com.BioTy.Planty.dto.chat;
 
 public class SendMessageRequestDto {
     private String message;
